Infrastructure/Major/WIP: Butcher mah boi and add a quasi build system
Kind of.  I just need a way of assembling this stuff so that it can run in JavaFX.  Loading the JS as ES6 modules was causing CORS-related failures in JFX.  This changes makes it so that you can no longer view the color picker through the Express server.  But you *can* now view it as a static HTML file.  So there's that.

It's janky.  It's ugly.  I need to come back and do this better later.  But, given the time constraints of the moment, this is what I'm going with.