Bugfix/Major: Make dragging knobs work again
And make it feel decent to use, too!

I tried reverting back to my old HTML Drag and Drop API code, but...
I ran into some issues in JavaFX that I wasn't able to work around.
For one, it was spitting out uncaught exceptions into the console (not
that I really care all that much).  But, more importantly, it was
setting the cursor to the "you can't drop things here" cursor, 100% of
the time while dragging.

I think both of these stemmed from the same issue.  It seems like
there's a bug in OpenJavaFX that causes it to not properly register the
start of DnD events.

I tried shimming in a thing where the JS 'dragstart' event would send a
faked mouse event to `Stage.onDragDetected`, and properly launch the
`startDragAndDrop` method, but... it was all to no avail.  Even when you
do that, it immediately terminates the drag, and keeps on doing what it
was doing before.

So, ultimately, we have to write our own DnD code.  Distasteful.

Of course, JFX doesn't even seem to have the `user-select` and `cursor:
grabbing` features implemented, so... who even cares about all this,
really!  SMH....