Refactor: Rethink handling of DOM-reading API
The main idea is really that we could fetch a particular element type
with generics, rather than explicit casts. The explicit casts are
annoying because they often require wrapping a statement in
parentheses, which then causes compilation errors about trying to
invoke the thing in parens as a function call, which (confusingly) is
something that you're supposed to fix by adding an explicit semicolon
at the end of the previous line.

Also in this commit: More-consistently using partial application.