UX: Hide parens in Advanced value preview
This makes it consistent with the new behavior in Simple, which is:
When using the `NLWord` output format, we never show parentheses in
the value preview, but, if it's a compound value (e.g. `green + 0.2`)
instead of a simple value (e.g. `green`), then add parentheses around
it when copying it to the clipboard.

I'm not at all in love with the idea of having them copy a different
thing that they see in the UI, but I think it's more accurate to their
intentions this way—I think!